feat: Move the transaction history loading logic to a custom hook
